解决ping工具一直占用CPU的问题

+++ network/ping/c++/main.cpp
@@ -52,6 +52,6 @@ int main(int argc, char *argv[])
 
     while (true)
     {
-        sleep(UINT32_MAX);
+        sleep(UINT16_MAX);
     }
 }
